Automatically backed up by Learn

+++ lib/create.sql
@@ -15,7 +15,7 @@ age INTEGER
 
 CREATE TABLE pledges (
 id INTEGER PRIMARY KEY,
-amount INTEGER,
+amount NUMERIC,
 user_id INTEGER,
 project_id INTEGER
 );
